Code cleanup and re-armoring
- Cleaned up and refactored a lot of code.
- Put the helper functions in its own file
- Added re-armoring, so we can add our comments/versions
- Fixed log message handling
- Put config specific stuff into its own file
- Updated go-mail, so we have access to embeds
- Added embedded file encryption
- Updated SPDX headers
