+++ go.mod
@@ -7,13 +7,10 @@ module github.com/wneessen/go-mail-middleware
 go 1.16
 
 require (
-	github.com/ProtonMail/gopenpgp/v2 v2.5.0
+	github.com/ProtonMail/gopenpgp/v2 v2.5.2
 	github.com/emersion/go-msgauth v0.6.6
-	github.com/wneessen/go-mail v0.3.9-0.20230128134908-3c0f5d03d844
+	github.com/wneessen/go-mail v0.3.9-0.20230129135243-cfaeb51bef7f
 	golang.org/x/text v0.6.0
 )
 
-require (
-	golang.org/x/crypto v0.1.0 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
-)
+require golang.org/x/crypto v0.1.0 // indirect
